Move sliding_window function from codecs module to common.py
