+++ tests/test_codecs.py
@@ -6,18 +6,6 @@ import pytest
 from RPLCD import codecs
 
 
-@pytest.mark.parametrize(['input_', 'lookahead', 'result'], [
-    ('hi', 0, [('h',), ('i',)]),
-    ('hi', 1, [('h', 'i'), ('i', ' ')]),
-    ('hi', 2, [('h', 'i', ' '), ('i', ' ', ' ')]),
-    ('', 0, []),
-    ('', 1, []),
-    ('', 7, []),
-])
-def test_window_function(input_, lookahead, result):
-    assert list(codecs._window(input_, lookahead)) == result
-
-
 @pytest.mark.parametrize(['input_', 'result_a00', 'result_a02'], [
     # Empty
     ('', [], []),
